--- src/components/reusable/rangeInput/index.ts
@@ -1,1 +1,0 @@
-export { RangeInput } from './rangeInput';
