--- src/components/reusable/rangeInput/rangeInput.scss
@@ -1,271 +1,0 @@
-@use '../../../common/scss/global.scss';
-@use '../../../common/scss/form-input.scss';
-@use '~@kyndryl-design-system/shidoka-foundation/scss/mixins/typography.scss';
-@use '~@kyndryl-design-system/shidoka-foundation/scss/mixins/spacing.scss';
-
-:host {
-  display: inline-block;
-}
-
-.slider {
-  -webkit-appearance: none;
-  width: 100%;
-  min-width: 323px;
-  height: 4px;
-  background: linear-gradient(
-    to right,
-    var(--kd-color-border-container-selected) 0%,
-    var(--kd-color-border-container-selected) calc(var(--val) * 1%),
-    #ccc calc(var(--val) * 1%),
-    #ccc 100%
-  );
-  border-radius: 4px;
-  outline: none;
-  transition: background 0.3s;
-  margin: 20px 0;
-}
-
-/* Track color when active */
-// .slider:focus {
-//   background: #ddd;
-// }
-
-/* WebKit (Chrome, Safari) Thumb */
-.slider::-webkit-slider-thumb {
-  -webkit-appearance: none;
-  appearance: none;
-  height: 24px;
-  width: 24px;
-  background: var(--kd-color-border-button-secondary-state-default);
-  border: 2px solid white;
-  border-radius: 50%;
-  cursor: pointer;
-  margin-top: -9px;
-  box-shadow: 0 0 0 2px white,
-    0 0 0 4px var(--kd-color-border-button-secondary-state-default);
-}
-
-/* Remove red ring by default, only on focus */
-.slider:not(:focus)::-webkit-slider-thumb {
-  box-shadow: none;
-}
-
-.slider::-webkit-slider-runnable-track {
-  height: 6px;
-  background: transparent;
-  border-radius: 3px;
-}
-
-.slider::-webkit-slider-progress {
-  background: var(--kd-color-border-container-selected);
-  height: 4px;
-  border-radius: 4px;
-}
-
-/* Firefox Thumb */
-.slider::-moz-range-thumb {
-  height: 20px;
-  width: 20px;
-  background: var(--kd-color-border-button-secondary-state-default);
-  border: 2px solid white;
-  border-radius: 50%;
-  cursor: pointer;
-  box-shadow: 0 0 0 4px white,
-    0 0 0 6px var(--kd-color-border-button-secondary-state-default);
-}
-
-/* Remove red ring on blur */
-.slider:not(:focus)::-moz-range-thumb {
-  box-shadow: none;
-}
-
-/* Firefox Track */
-.slider::-moz-range-track {
-  background: #ccc;
-  height: 4px;
-  border-radius: 4px;
-}
-
-/* Progress color (Firefox only) */
-.slider::-moz-range-progress {
-  background: var(--kd-color-border-container-selected);
-  height: 4px;
-  border-radius: 4px;
-}
-
-/* Disabled state */
-.slider:disabled {
-  opacity: 0.5;
-  cursor: not-allowed;
-}
-
-// .slider-tooltip {
-//   @include typography.type-body-02;
-//   position: absolute;
-//   top: -1.8em;
-//   left: 50%;
-//   padding: var(--kd-spacing-4) var(--kd-spacing-8);
-//   background: var(--kd-color-background-ui-default-dark);
-//   color: var(--kd-color-text-variant-inversed);
-//   border-radius: var(--kd-spacing-4);
-//   height: 2em;
-//   min-width: var(--kd-spacing-40);
-//   text-align: center;
-//   white-space: nowrap;
-//   visibility: visible;
-//   z-index: 2;
-//   transform-origin: center bottom;
-//   transition: visibility 150ms ease-out, transform 150ms ease-out;
-//   transform: translateX(-50%);
-
-//   /* Add the arrow before the tooltip */
-//   &::before {
-//     content: '';
-//     position: absolute;
-//     top: 100%;
-//     left: 50%;
-//     width: 0;
-//     height: 0;
-//     transform: translateX(-50%);
-//     border-left: var(--kd-spacing-4) solid transparent; /* Left side of the arrow */
-//     border-right: var(--kd-spacing-4) solid transparent; /* Right side of the arrow */
-//     border-top: var(--kd-spacing-4) solid
-//       var(--kd-color-background-ui-default-dark); /* Bottom side of the arrow */
-//   }
-// }
-
-// input[type='range'] {
-//   display: block;
-//   width: 100%;
-//   -webkit-appearance: none;
-//   appearance: none;
-//   cursor: pointer;
-//   outline: none;
-//   min-width: 452px;
-//   height: 4px;
-//   margin: 0px;
-//   border-radius: var(--kd-spacing-4);
-//   background-color: var(--kd-color-border-level-tertiary);
-// }
-
-// input[type='range']::-webkit-slider-thumb {
-//   -webkit-appearance: none;
-//   appearance: none;
-//   height: 16px;
-//   width: 16px;
-//   background-color: var(--kd-color-border-button-secondary-state-default);
-//   border-radius: 50%;
-//   box-shadow: -2px 0 0 0 var(--kd-color-border-variants-inverse),
-//     var(--kd-spacing-2) 0 0 0 var(--kd-color-border-variants-inverse);
-//   outline: var(--kd-spacing-2) solid transparent;
-//   outline-offset: var(--kd-spacing-2);
-//   transition: 0.2s ease-in-out;
-// }
-
-// input[type='range']:not(:disabled)::-webkit-slider-thumb:hover {
-//   height: 24px;
-//   width: 24px;
-// }
-// input[type='range']:not(:disabled):active::-webkit-slider-thumb {
-//   height: 24px;
-//   width: 24px;
-//   box-shadow: 0 0 0 2px var(--kd-color-border-variants-inverse),
-//     0 0 0 4px var(--kd-color-border-variants-focus);
-// }
-// input[type='range']:not(:disabled):focus::-webkit-slider-thumb {
-//   height: 24px;
-//   width: 24px;
-//   box-shadow: 0 0 0 2px var(--kd-color-border-variants-inverse),
-//     0 0 0 4px var(--kd-color-border-variants-focus);
-// }
-
-// input[type='range']::-moz-range-thumb {
-//   height: var(--kd-spacing-12);
-//   width: var(--kd-spacing-12);
-//   background-color: var(--kd-color-border-button-secondary-state-default);
-//   border-radius: 50%;
-//   box-shadow: -2px 0 0 0 var(--kd-color-border-variants-inverse),
-//     var(--kd-spacing-2) 0 0 0 var(--kd-color-border-variants-inverse);
-//   outline: var(--kd-spacing-2) solid transparent;
-//   outline-offset: var(--kd-spacing-2);
-//   transition: 0.2s ease-in-out;
-// }
-
-// input[type='number'] {
-//   -webkit-appearance: none;
-//   appearance: none;
-//   display: block;
-//   min-width: var(--kd-spacing-40);
-//   border-width: 1px;
-//   border-style: solid;
-//   border-radius: var(--kd-spacing-4);
-//   height: var(--kd-spacing-32);
-//   padding: var(--kd-spacing-0) var(--kd-spacing-8);
-// }
-
-// .number-input {
-//   display: flex;
-//   gap: var(--kd-spacing-8);
-//   align-items: center;
-// }
-
-// .steps {
-//   height: 0.875rem;
-//   font-size: 0.875rem;
-//   line-height: 1;
-//   margin-bottom: 25px;
-// }
-
-// .step-container {
-//   position: absolute;
-//   inset-block-start: 0;
-//   inset-inline-start: 0%;
-//   content: '';
-// }
-
-// .step-end {
-//   position: absolute;
-//   inset-block-start: 0;
-//   inset-inline-start: 100%;
-//   content: '';
-// }
-
-// .step-label {
-//   @include typography.type-body-02;
-//   transform: translateX(0);
-//   position: absolute;
-//   inset-block-start: calc(2rem + 0.25rem);
-//   word-break: normal;
-//   padding-top: 10px;
-//   width: 1px;
-//   height: 10px;
-//   display: flex;
-//   justify-content: center;
-//   background-color: var(--kd-color-border-level-primary);
-// }
-
-// .step-label-end {
-//   @include typography.type-body-02;
-//   padding-top: 10px;
-//   width: 1px;
-//   height: 10px;
-//   display: flex;
-//   justify-content: center;
-//   background-color: var(--kd-color-border-level-primary);
-//   transform: translateX(-100%);
-//   position: absolute;
-//   inset-block-start: calc(2rem + 0.25rem);
-//   word-break: normal;
-// }
-
-// .tick {
-//   position: absolute;
-//   width: var(--kd-spacing-4);
-//   background: var(--kd-color-background-accent-secondary);
-//   height: 5px;
-//   transform: translateX(-50%);
-//   margin-top: -4px;
-//   &:first-child {
-//     visibility: hidden;
-//   }
-// }
