--- src/components/reusable/sliderInput1/index.ts
@@ -1,1 +1,0 @@
-export { SliderInput1 } from './sliderInput1';
