--- src/components/reusable/sliderInput1/sliderInput1.scss
@@ -1,297 +1,0 @@
-@use '../../../common/scss/global.scss';
-@use '../../../common/scss/form-input.scss';
-@use '~@kyndryl-design-system/shidoka-foundation/scss/mixins/typography.scss';
-@use '~@kyndryl-design-system/shidoka-foundation/scss/mixins/spacing.scss';
-
-:host {
-  display: inline-block;
-}
-
-.slider-wrapper {
-  position: relative;
-  flex-grow: 1;
-}
-
-.slider-container {
-  padding-block-start: 0;
-  padding-block-end: 10px;
-}
-
-.margin-bottom20 {
-  margin-bottom: 20px;
-}
-
-.slider-tooltip {
-  @include typography.type-body-02;
-  position: absolute;
-  top: -27px;
-  left: 50%;
-  padding: var(--kd-spacing-4) var(--kd-spacing-8);
-  background: var(--kd-color-background-ui-default-dark);
-  color: var(--kd-color-text-variant-inversed);
-  border-radius: var(--kd-spacing-4);
-  height: 2em;
-  min-width: var(--kd-spacing-40);
-  text-align: center;
-  white-space: nowrap;
-  visibility: visible;
-  z-index: 2;
-  transform-origin: center bottom;
-  transition: visibility 150ms ease-out, transform 150ms ease-out;
-  transform: translateX(-50%);
-
-  /* Add the arrow before the tooltip */
-  &::before {
-    content: '';
-    position: absolute;
-    top: 100%;
-    left: 50%;
-    width: 0;
-    height: 0;
-    transform: translateX(-50%);
-    border-left: var(--kd-spacing-4) solid transparent; /* Left side of the arrow */
-    border-right: var(--kd-spacing-4) solid transparent; /* Right side of the arrow */
-    border-top: var(--kd-spacing-4) solid
-      var(--kd-color-background-ui-default-dark); /* Bottom side of the arrow */
-  }
-}
-
-input[type='range'] {
-  width: 100%;
-  min-width: 323px;
-  overflow: visible;
-  padding: calc(var(--thumb-height, 24px) / 2) 0;
-}
-
-input[type='range']:disabled {
-  background: none;
-  cursor: not-allowed;
-}
-
-input[type='range'] {
-  color: var(--kd-color-border-container-selected);
-  position: relative;
-  margin: 0px;
-  --thumb-height: 20px;
-  --thumb-width: 20px;
-  --track-height: 4px;
-  --track-color: var(--kd-color-border-level-tertiary);
-  --clip-edges: 0.125em;
-}
-
-// input[type='range']:disabled {
-//   filter: grayscale(1);
-//   opacity: 0.3;
-//   cursor: not-allowed;
-// }
-
-input[type='range'],
-input[type='range']::-webkit-slider-runnable-track {
-  -webkit-appearance: none;
-  transition: all ease 100ms;
-}
-
-input[type='range']::-webkit-slider-runnable-track,
-input[type='range']::-webkit-slider-thumb {
-  position: relative;
-}
-
-input[type='range']::-webkit-slider-thumb {
-  -webkit-appearance: none;
-  transition: all ease 100ms;
-  height: var(--thumb-height);
-  --thumb-radius: calc((var(--thumb-height) * 0.5) - 1px);
-  --clip-top: calc((var(--thumb-height) - var(--track-height)) * 0.5 - 0.5px);
-  --clip-bottom: calc(var(--thumb-height) - var(--clip-top));
-  --clip-further: calc(100% + 4px);
-  --box-fill: calc(-100vmax - var(--thumb-width, var(--thumb-height))) 0 0
-    100vmax currentColor;
-
-  width: var(--thumb-width, var(--thumb-height));
-  background-color: var(--kd-color-border-button-secondary-state-default);
-  box-shadow: -2px 0 0 0 var(--kd-color-border-variants-inverse),
-    var(--box-fill), 2px 0 0 0 var(--kd-color-border-variants-inverse);
-
-  border-radius: var(--thumb-width, var(--thumb-height));
-
-  clip-path: polygon(
-    100% -1px,
-    var(--clip-edges) -1px,
-    0 var(--clip-top),
-    -100vmax var(--clip-top),
-    -100vmax var(--clip-bottom),
-    0 var(--clip-bottom),
-    var(--clip-edges) 100%,
-    var(--clip-further) var(--clip-further)
-  );
-  //   outline: none;
-  transition: all 0.2s ease-in-out;
-}
-
-input[type='range']:hover::-webkit-slider-thumb {
-  cursor: grab;
-}
-
-input[type='range']:active::-webkit-slider-thumb {
-  cursor: grabbing;
-}
-
-input[type='range']::-webkit-slider-runnable-track {
-  overflow: hidden;
-  background: linear-gradient(var(--track-color) 0 0) scroll no-repeat center /
-    100% calc(var(--track-height) + 1px);
-  border-radius: 4px;
-}
-
-/* For webkit disabled thumb */
-input[type='range']:disabled::-webkit-slider-thumb {
-  background: var(--kd-color-text-level-disabled);
-  cursor: not-allowed;
-}
-
-/* For firefox disabled thumb */
-input[type='range']:disabled::-moz-range-thumb:hover {
-  background: var(--kd-color-text-level-disabled);
-  cursor: not-allowed;
-}
-
-input[type='range'],
-input[type='range']::-moz-range-track,
-input[type='range']::-moz-range-thumb {
-  appearance: none;
-  transition: all ease 100ms;
-}
-
-input[type='range']::-moz-range-track,
-input[type='range']::-moz-range-thumb,
-input[type='range']::-moz-range-progress {
-  height: 8px;
-  background: var(--kd-color-border-button-secondary-state-default);
-}
-
-input[type='range']::-moz-range-thumb {
-  height: 20px;
-  border: 0;
-  width: 20px;
-  border-radius: 50%;
-  background-color: var(--kd-color-border-button-secondary-state-default);
-  cursor: grab;
-  box-shadow: -2px 0 0 0 var(--kd-color-border-variants-inverse),
-    var(--kd-spacing-2) 0 0 0 var(--kd-color-border-variants-inverse);
-}
-
-input[type='range']:active::-moz-range-thumb {
-  cursor: grabbing;
-}
-
-input[type='range']::-moz-range-track {
-  width: 100%;
-  height: 4px;
-  background: var(--track-color);
-  border-radius: 4px;
-}
-
-input[type='range']::-moz-range-progress {
-  appearance: none;
-  background: currentColor;
-  transition-delay: 30ms;
-}
-
-input[type='range']::-moz-range-track,
-input[type='range']::-moz-range-progress {
-  height: 4px;
-  border-radius: var(--track-height);
-}
-
-input[type='range']:disabled::-webkit-slider-thumb {
-  cursor: not-allowed;
-}
-
-input[type='range']:disabled::-moz-range-thumb {
-  cursor: not-allowed;
-}
-
-input[type='number'] {
-  -webkit-appearance: none;
-  appearance: none;
-  display: block;
-  min-width: var(--kd-spacing-40);
-  border-width: 1px;
-  border-style: solid;
-  border-radius: var(--kd-spacing-4);
-  height: var(--kd-spacing-32);
-  text-align: center;
-}
-
-.number-input {
-  margin-inline-start: 15px;
-  padding-block-start: 0.2rem;
-}
-
-.tick {
-  position: absolute;
-  --thumb-height: 20px; /* Thumb height */
-  --track-height: 5px; /* Track height */
-  width: var(--kd-spacing-4);
-  background: var(--kd-color-background-accent-secondary);
-  height: 5px;
-  transform: translateX(-50%);
-  top: calc(var(--thumb-height) - var(--track-height) + 2.6px);
-}
-
-.tick-label {
-  @include typography.type-body-02;
-  position: absolute;
-  top: 100%;
-  left: 50%;
-  transform: translateX(-50%);
-  margin-top: -12px;
-
-  /* Add a vertical line above the label */
-  &::before {
-    content: '';
-    position: absolute;
-    top: -11px;
-    left: 50%;
-    transform: translateX(-50%);
-    width: 1px;
-    height: 10px;
-    background-color: var(--kd-color-border-level-primary);
-  }
-}
-
-.tick-container {
-  margin-top: -14px;
-}
-
-.tick-container :last-child {
-  .tick-label {
-    transform: translateX(-100%);
-  }
-}
-
-.step-label {
-  @include typography.type-body-02;
-  transform: translateX(0);
-  position: absolute;
-  inset-block-start: 2rem;
-  word-break: normal;
-  padding-top: 10px;
-  width: 1px;
-  height: 8px;
-  display: flex;
-  justify-content: center;
-  background-color: var(--kd-color-border-level-primary);
-}
-
-/* Hide native arrow controls */
-input::-webkit-outer-spin-button,
-input::-webkit-inner-spin-button {
-  -webkit-appearance: none;
-  margin: 0;
-}
-
-input[type='number'] {
-  -moz-appearance: textfield;
-  /* Firefox */
-}
